Rename base to utils, add extendInstance, add RequiredKeys
