--- packages/drift/src/base/types.ts
@@ -1,7 +1,0 @@
-export type EmptyObject = Record<string, never>;
-
-/**
- * Combines members of an intersection into a readable type.
- * @see https://twitter.com/mattpocockuk/status/1622730173446557697?s=20&t=NdpAcmEFXY01xkqU3KO0Mg
- */
-export type Prettify<T> = { [K in keyof T]: T[K] } & unknown;
