+++ packages/drift/src/contract/factories/createCachedReadContract.test.ts
@@ -1,7 +1,7 @@
-import { IERC20 } from "src/base/testing/IERC20";
-import { ALICE, BOB } from "src/base/testing/accounts";
 import { ReadContractStub } from "src/contract/stubs/ReadContractStub";
 import type { Event } from "src/contract/types/Event";
+import { IERC20 } from "src/utils/testing/IERC20";
+import { ALICE, BOB } from "src/utils/testing/accounts";
 import { describe, expect, it } from "vitest";
 import { createCachedReadContract } from "./createCachedReadContract";
